ATO-957: Configure environment variables
Private and public keys will be needed for decryption and signature validation.

The local private key has been commited deliberately, it is not a mistake. The key pair was generated fresh and will only be used for testing, both locally and as part of the pre-merge GitHub workflow.

In deployed envs, the private key will be retrieved from Secrets Manager, and the public key from Parameter Store. This key pair is different from the one which has been commited here.