First pass at a SPOT stub
Currently just responds with a hard-coded identity. This will need to change over time.
This also does no validation of the input and does not serve to test that the integration with SPOT is full functional. It simply allows us to exercise the IPV Core stub.

Deployment at this stage is manual. A pipeline is to follow.