BAU: Adds conditional env var to template
SAM local does not provide environemt variables
to the lambda functions unless they are specified
in the cloudformation template[1]. This is annoying
when trying to pass the local stack endpoint when locally
running the lambda.

A workaround is to make this confitional on being a local
environment, and use AWS::NoValue to not set this in a deployed
environment. It's not ideal but SAM local forces us to do this

[1] - https://github.com/aws/aws-sam-cli/issues/139#issuecomment-334977285