ATO-1077: Handle missing form fields
As we plan to render the different claim fields
optionally based on the request claims. We need to
be able to handle the form fields being undefined
or empty submissions.

This updates the mapFormToUserIdentity function
and to validate that a coreIdentity payload has
been submitted and handle missing other fields.
Also adds a helper function to display nicer
JSON error messages