ATO-1795: Add hidden styling and script input
I would like to add the ability of the IPV stub to return specifc
oAuth errors back to orchestration. This will allow us to test the
specific session_invalidated case we plan to add. To do this I would
like to be able to hide elements based on a drop down selection,
this required a hidden styling class and the ability to add client
side js to our rendering of the page. This adds both of these